// src/app/models/yarnDeliverySearchResult.ts
export interface YarnDeliverySearchResult {
  dcDtsId: number;
  orderDtsId: number;
  shadeId: number;
  shadeName: string;
  blendId: number;
  blendName: string;
  countsId: number;
  counts: string;
  lot: string;
  hsnCode: number;
  orderQuantity: number;
  deliveredQuantity: number;
  previousDeliveredQuantity: number;
  dcId: number;
  dcNo: string;
  dcDate: string;
  partyId: number;
  partyName: string;
}